Refactor TaskListTypes Class
